send_template_email utility function
rethinking mixin approach based on review
